Исправлены формула аннуитетного кредита, ссылка на страницу с графиком, добавлена инициализация кредитных программ

+++ WebBank/AppCore/Entities/CreditProgram.cs
@@ -6,7 +6,7 @@ public class CreditProgram
 {
     public int Id { get; set; }
     public required string Name { get; set; }
-    public int Percent { get; set; }
+    public decimal Percent { get; set; }
     public int Period { get; set; }
 
     [ForeignKey(nameof(Currency))]
